PJ: simplify link recipes
DOSBox does not implement "COPY x+y x" correctly, where x appears as
both source and destination.  This was used in some make rules to
concatenate several files into a single response file for wlib.

There seems to be no reason not to use multiple response file arguments
to wlib instead.  Perhaps this was not supported in earlier versions?